moved reducers file to helpers folder
