+++ src/components/ViewSnippet/ViewSnippet.jsx
@@ -1,5 +1,5 @@
 import React, { useReducer } from "react";
-import { ACTIONS, commentReducer } from "../App/reducers";
+import { ACTIONS, commentReducer } from "../../helpers/reducers";
 // Import used to format / check js code
 import prettier from "https://unpkg.com/prettier@2.8.0/esm/standalone.mjs";
 import parserBabel from "https://unpkg.com/prettier@2.8.0/esm/parser-babel.mjs";
